AEM Page persistence strategy: Saving of Context-Aware configuration collections - restore previous behavior of setting last modified date of all items regardless of actual data changes or not. (#10)
Introduce a new OSGi configuration property collectionMarkAllItemsUpdated attached to the backward-compatible behavior set to true by default.
If set to false, the new behavior of is applied. But this new behavior with updating last modified only on the affected items seems to create problems with publishing in AEMaaCS cloud instances.